+++ score-multimodule/src/commonMain/kotlin/com/kjipo/score/Transform.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Translation(
-    val xShift: Int, val yShift: Int
+    val xShift: Double, val yShift: Double
 )
